% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{mean}
\alias{mean}
\alias{mean,sparse_numeric-method}
\title{Mean of sparse vector}
\usage{
\S4method{mean}{sparse_numeric}(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}
}
\value{
A numeric mean value.
}
\description{
Generic and method to compute the mean of a \code{sparse_numeric} vector.
}
