% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Norm of sparse vector}
\usage{
norm(object)

\S4method{norm}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\value{
A numeric norm value.
}
\description{
Generic and method to compute the Euclidean norm of a \code{sparse_numeric} vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{norm(sparse_numeric)}: Compute norm for sparse_numeric.

}}
