% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{plot}
\alias{plot}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot sparse vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{First sparse vector.}

\item{y}{Second sparse vector.}
}
\value{
A plot object.
}
\description{
Creates a dot plot showing where two sparse vectors overlap.
}
