% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{setAs}
\alias{setAs}
\title{Convert dense to sparse numeric}
\arguments{
\item{from}{A \code{sparse_numeric} object.}
}
\value{
A \code{sparse_numeric} object.

A numeric vector.
}
\description{
Converts a dense numeric vector to a \code{sparse_numeric} object.

Converts a \code{sparse_numeric} object to a regular numeric vector.
}
