% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{show}
\alias{show}
\alias{show,sparse_numeric-method}
\title{Show a sparse vector}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\value{
Printed output to console.
}
\description{
Prints a readable representation of a \code{sparse_numeric} object.
}
