% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{sort}
\alias{sort}
\alias{sort,sparse_numeric-method}
\title{Sort sparse vector}
\usage{
\S4method{sort}{sparse_numeric}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{decreasing}{Logical; if TRUE, sort in decreasing order.}

\item{...}{additional arguments passed to or from other methods}
}
\value{
A sorted \code{sparse_numeric} object.
}
\description{
Sorts values in the sparse vector by position.
}
