% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Add two sparse vectors}
\usage{
sparse_add(x, y)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{First sparse vector.}

\item{y}{Second sparse vector.}
}
\value{
A \code{sparse_numeric} object representing the sum.
}
\description{
Generic and method to add two \code{sparse_numeric} vectors element-wise.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_add(x = sparse_numeric, y = sparse_numeric)}: Compute the addtion vector for two sparse_numeric vectors.

}}
