% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Cross product of two sparse vectors}
\usage{
sparse_crossprod(x, y)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{First sparse vector.}

\item{y}{Second sparse vector.}
}
\value{
A numeric value.
}
\description{
Computes the cross product (dot product) of two \code{sparse_numeric} vectors.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_crossprod(x = sparse_numeric, y = sparse_numeric)}: Compute the cross product for two sparse_numeric vectors.

}}
