% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse vectors}
\usage{
sparse_mult(x, y)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{First sparse vector.}

\item{y}{Second sparse vector.}
}
\value{
A \code{sparse_numeric} object.
}
\description{
Generic and method to multiply two \code{sparse_numeric} vectors element-wise.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_mult(x = sparse_numeric, y = sparse_numeric)}: Compute the multplication vector for two sparse_numeric vectors.

}}
