% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Numeric Class}
\value{
A \code{sparse_numeric} object.
}
\description{
Represents a sparse numeric vector with values, positions, and total length.
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of nonzero entries.}

\item{\code{pos}}{Integer vector indicating positions of nonzero entries.}

\item{\code{length}}{Integer specifying the overall vector length.}
}}

