% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract two sparse vectors}
\usage{
sparse_sub(x, y)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{First sparse vector.}

\item{y}{Second sparse vector.}
}
\value{
A \code{sparse_numeric} object.
}
\description{
Generic and method to subtract one \code{sparse_numeric} vector from another.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_sub(x = sparse_numeric, y = sparse_numeric)}: Compute the subtraction vector of two sparse_numeric vectors.

}}
