% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{sparse_sum}
\alias{sparse_sum}
\alias{sparse_sum,sparse_numeric-method}
\title{Sum of sparse vector}
\usage{
sparse_sum(object)

\S4method{sparse_sum}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\value{
A numeric value.
}
\description{
Computes the sum of all nonzero values in a \code{sparse_numeric} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_sum(sparse_numeric)}: Compute the sum of values for sparse_numeric.

}}
