% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize sparse vector}
\usage{
standardize(object)

\S4method{standardize}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\value{
A normalized \code{sparse_numeric} object.
}
\description{
Generic and method to normalize a \code{sparse_numeric} vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{standardize(sparse_numeric)}: Compute the standardize sparse vector for sparse_numeric.

}}
