% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{validation}
\alias{validation}
\title{Validate a sparse_numeric object}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\value{
TRUE if valid; otherwise returns an error message.
}
\description{
Checks if the sparse numeric object meets the basic requirements of being valid.
}
