% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Code.R
\name{mean}
\alias{mean}
\alias{mean,sparse_numeric-method}
\title{Mean of sparse vector}
\usage{
mean(object)

\S4method{mean}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\value{
A numeric mean value.
}
\description{
Generic and method to compute the mean of a \code{sparse_numeric} vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{mean(sparse_numeric)}: Compute mean for sparse_numeric.

}}
